refact : Friend Add 수정 - Script 통일
